Linker Errors with MGS3 Codebase
Pausing Development while I continue studying Beginning CPP through Game Programming.
-------------------------------------------------------------------------------------------
1>CppCodeSampleCollection.obj : error LNK2019: unresolved external symbol "public: class Item & __thiscall Item::operator=(class Item const &)" (??4Item@@QAEAAV0@ABV0@@Z) referenced in function _main
1>HUD.obj : error LNK2001: unresolved external symbol "public: class Item & __thiscall Item::operator=(class Item const &)" (??4Item@@QAEAAV0@ABV0@@Z)
1>Inventory.obj : error LNK2001: unresolved external symbol "public: class Item & __thiscall Item::operator=(class Item const &)" (??4Item@@QAEAAV0@ABV0@@Z)
1>CppCodeSampleCollection.obj : error LNK2019: unresolved external symbol "public: class Player && __thiscall Player::operator=(class Player const &)" (??4Player@@QAE$$QAV0@ABV0@@Z) referenced in function _main
1>CppCodeSampleCollection.obj : error LNK2019: unresolved external symbol "public: class Inventory * __thiscall Player::GetInventory(void)" (?GetInventory@Player@@QAEPAVInventory@@XZ) referenced in function _main
1>D:\Documents\GitHub\CppCodeSampleCollection\CppCodeSampleRepo\CppCodeSampleCollection\Debug\CppCodeSampleCollection.exe : fatal error LNK1120: 3 unresolved externals
1>Done building project "CppCodeSampleCollection.vcxproj" -- FAILED.

+++ CppCodeSampleCollection/Food.cpp
@@ -38,14 +38,14 @@ int Food::GetAge() {
 	return _age;
 }
 
-bool Food::SetAge(float value) {
+bool Food::SetAge(int value) {
 	_age = value;
 	if (_age > SPOILING_AGE) {
 		return true;
 	}
 	return false;
 }
-bool Food::IncrementAge(float value) {
+bool Food::IncrementAge(int value) {
 	SetAge(_age - value);
 	if (_age > SPOILING_AGE) {
 		return true;
